LayoutTests/fast/events/onload-webkit-before-webcore.html

<script>layoutTestController.dumpAsText()</script>
<BODY onload="alert('This should be the first line.')">
This should be the second line. The first line should be an ALERT.
    <iframe src="resources/aFileThatDoesNotExist"></iframe>
    <!--
        This tests for a regression against
        https://bugs.webkit.org/show_bug.cgi?id=7693
        WebKit relies on unpredictable timing for onload events
        
        The body load event should be dispatched before the
        WebFrame load delegate is sent the webView:didFinishLoadForFrame: message.
    -->
</BODY>
