third_party/blink/web_tests/platform/generic/fast/events/onload-webkit-before-webcore-expected.txt

ALERT: This should be the first line.
This should be the second line. The first line should be an ALERT. 
